feat(entity): 회원 관련된 엔티티들 추가

+++ src/main/java/com/nhnacademy/back/account/socialauth/domain/entity/SocialAuth.java
@@ -13,7 +13,7 @@ public class SocialAuth {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int socialAuthId;
+	private long socialAuthId;
 
 	@Column(nullable = false)
 	@Enumerated(EnumType.STRING)
